admin/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Already in the gateway venv — shared with gateway package
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
redis[hiredis]>=5.0.0
PyJWT>=2.8.0
